+++ src/plugins/workspace/public/plugin.test.ts
@@ -3,10 +3,26 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { coreMock } from '../../../core/public/mocks';
+import { workspaceClientMock, WorkspaceClientMock } from './workspace_client.mock';
+import { chromeServiceMock, coreMock } from '../../../core/public/mocks';
 import { WorkspacePlugin } from './plugin';
 
 describe('Workspace plugin', () => {
+  const getSetupMock = () => ({
+    ...coreMock.createSetup(),
+    chrome: chromeServiceMock.createSetupContract(),
+  });
+  beforeEach(() => {
+    WorkspaceClientMock.mockClear();
+    Object.values(workspaceClientMock).forEach((item) => item.mockClear());
+  });
+  it('#setup', async () => {
+    const setupMock = getSetupMock();
+    const workspacePlugin = new WorkspacePlugin();
+    await workspacePlugin.setup(setupMock);
+    expect(WorkspaceClientMock).toBeCalledTimes(1);
+  });
+
   it('#call savedObjectsClient.setCurrentWorkspace when current workspace id changed', () => {
     const workspacePlugin = new WorkspacePlugin();
     const coreStart = coreMock.createStart();